migrate to error macros

+++ src/computer_shader_task.h
@@ -66,7 +66,7 @@ private:
 	[[nodiscard]] RID _get_sampler(RenderingDevice::SamplerFilter filter, RenderingDevice::SamplerRepeatMode repeat_mode) const;
 	[[nodiscard]] Variant _get_default_uniform(RenderingDevice::UniformType type, Dictionary user_attributes) const;
 	Ref<RDUniformBuffer> _get_uniform_buffer(int64_t binding, int64_t set);
-	void _set_uniform_buffer(int64_t binding, int64_t set, const RID& buffer_rid);
+	void _set_uniform_buffer(int32_t binding, int32_t set, const RID& buffer_rid);
 	void _update_buffers(int64_t kernel_index);
 	void _bind_uniform_sets(int64_t kernel_index, int64_t compute_list, RenderingDevice* rd);
 	void _dispatch(int64_t kernel_index, Vector3i thread_groups);
